Split PostHog analytics tests
