+++ cmuxTests/PostHogAnalyticsPropertiesTests.swift
@@ -1,0 +1,141 @@
+import Foundation
+import Testing
+
+#if canImport(cmux_DEV)
+@testable import cmux_DEV
+
+@Suite(.serialized)
+struct PostHogAnalyticsPropertiesTests {
+    @Test
+    func dailyActivePropertiesIncludeVersionAndBuild() {
+        let properties = PostHogAnalytics.dailyActiveProperties(
+            dayUTC: "2026-02-21",
+            reason: "didBecomeActive",
+            infoDictionary: [
+                "CFBundleShortVersionString": "0.31.0",
+                "CFBundleVersion": "230",
+            ]
+        )
+
+        #expect(properties["day_utc"] as? String == "2026-02-21")
+        #expect(properties["reason"] as? String == "didBecomeActive")
+        #expect(properties["app_version"] as? String == "0.31.0")
+        #expect(properties["app_build"] as? String == "230")
+    }
+
+    @Test
+    func superPropertiesIncludePlatformVersionAndBuild() {
+        let properties = PostHogAnalytics.superProperties(
+            infoDictionary: [
+                "CFBundleShortVersionString": "0.31.0",
+                "CFBundleVersion": "230",
+            ]
+        )
+
+        #expect(properties["platform"] as? String == "cmuxterm")
+        #expect(properties["app_version"] as? String == "0.31.0")
+        #expect(properties["app_build"] as? String == "230")
+    }
+
+    @Test
+    func hourlyActivePropertiesIncludeVersionAndBuild() {
+        let properties = PostHogAnalytics.hourlyActiveProperties(
+            hourUTC: "2026-02-21T14",
+            reason: "didBecomeActive",
+            infoDictionary: [
+                "CFBundleShortVersionString": "0.31.0",
+                "CFBundleVersion": "230",
+            ]
+        )
+
+        #expect(properties["hour_utc"] as? String == "2026-02-21T14")
+        #expect(properties["reason"] as? String == "didBecomeActive")
+        #expect(properties["app_version"] as? String == "0.31.0")
+        #expect(properties["app_build"] as? String == "230")
+    }
+
+    @Test
+    func hourlyPropertiesOmitVersionFieldsWhenUnavailable() {
+        let properties = PostHogAnalytics.hourlyActiveProperties(
+            hourUTC: "2026-02-21T14",
+            reason: "activeTimer",
+            infoDictionary: [:]
+        )
+
+        #expect(properties["hour_utc"] as? String == "2026-02-21T14")
+        #expect(properties["reason"] as? String == "activeTimer")
+        #expect(properties["app_version"] == nil)
+        #expect(properties["app_build"] == nil)
+    }
+
+    @Test
+    func propertiesOmitVersionFieldsWhenUnavailable() {
+        let superProperties = PostHogAnalytics.superProperties(infoDictionary: [:])
+        #expect(superProperties["platform"] as? String == "cmuxterm")
+        #expect(superProperties["app_version"] == nil)
+        #expect(superProperties["app_build"] == nil)
+
+        let dailyProperties = PostHogAnalytics.dailyActiveProperties(
+            dayUTC: "2026-02-21",
+            reason: "activeTimer",
+            infoDictionary: [:]
+        )
+        #expect(dailyProperties["day_utc"] as? String == "2026-02-21")
+        #expect(dailyProperties["reason"] as? String == "activeTimer")
+        #expect(dailyProperties["app_version"] == nil)
+        #expect(dailyProperties["app_build"] == nil)
+    }
+
+    @Test
+    func flushPolicyIncludesDailyAndHourlyActiveEvents() {
+        #expect(PostHogAnalytics.shouldFlushAfterCapture(event: "cmux_daily_active"))
+        #expect(PostHogAnalytics.shouldFlushAfterCapture(event: "cmux_hourly_active"))
+        #expect(!PostHogAnalytics.shouldFlushAfterCapture(event: "cmux_other_event"))
+    }
+
+    @Test
+    func activeEventCaptureFlushesBeforeShutdown() throws {
+        let suiteName = "cmux.posthog.analytics.tests.\(UUID().uuidString)"
+        let defaults = try #require(UserDefaults(suiteName: suiteName))
+        defer {
+            defaults.removePersistentDomain(forName: suiteName)
+        }
+
+        let fixedDate = try #require(Calendar(identifier: .iso8601).date(from: DateComponents(
+            timeZone: TimeZone(secondsFromGMT: 0),
+            year: 2026,
+            month: 2,
+            day: 21,
+            hour: 14
+        )))
+        let capturedQueue = DispatchQueue(label: "com.cmux.tests.posthog.capture")
+        var capturedEvents: [(event: String, properties: [String: Any])] = []
+        let flushCalled = DispatchSemaphore(value: 0)
+        let analytics = PostHogAnalytics.makeForTesting(
+            workQueue: DispatchQueue(label: "com.cmux.tests.posthog.analytics"),
+            didStart: true,
+            userDefaults: defaults,
+            now: { fixedDate },
+            capturePostHog: { event, properties in
+                capturedQueue.sync {
+                    capturedEvents.append((event: event, properties: properties))
+                }
+            },
+            flushPostHog: {
+                flushCalled.signal()
+            }
+        )
+
+        analytics.trackActive(reason: "didBecomeActive")
+        #expect(flushCalled.wait(timeout: .now() + .seconds(1)) == .success)
+        let events = capturedQueue.sync { capturedEvents }
+        #expect(events.map(\.event) == ["cmux_daily_active", "cmux_hourly_active"])
+        let dailyEvent = try #require(events.first)
+        let hourlyEvent = try #require(events.dropFirst().first)
+        #expect(dailyEvent.properties["day_utc"] as? String == "2026-02-21")
+        #expect(dailyEvent.properties["reason"] as? String == "didBecomeActive")
+        #expect(hourlyEvent.properties["hour_utc"] as? String == "2026-02-21T14")
+        #expect(hourlyEvent.properties["reason"] as? String == "didBecomeActive")
+    }
+}
+#endif
